# Speed up vegm file reading