--- a/pftools/python/requirements.txt
+++ b/pftools/python/requirements.txt
@@ -1,5 +1,6 @@
 PyYAML==5.4
 xarray
+pandas
 numba
 numpy
 dask